Split out SpecBlock, SpecBlockAssocition and ScriptTemplateAssociation into seperate files
