+++ src/lsst/cmservice/db/campaign.py
@@ -15,7 +15,9 @@ from .dbid import DbId
 from .element import ElementMixin
 from .enums import SqlStatusEnum
 from .production import Production
-from .specification import SpecBlock, SpecBlockAssociation, Specification
+from .spec_block import SpecBlock
+from .spec_block_association import SpecBlockAssociation
+from .specification import Specification
 
 if TYPE_CHECKING:
     from .script import Script
