+++ src/lsst/cmservice/db/job.py
@@ -14,7 +14,9 @@ from .dbid import DbId
 from .element import ElementMixin
 from .enums import SqlStatusEnum
 from .group import Group
-from .specification import SpecBlock, SpecBlockAssociation, Specification
+from .spec_block import SpecBlock
+from .spec_block_association import SpecBlockAssociation
+from .specification import Specification
 from .step import Step
 
 if TYPE_CHECKING:
