+++ src/lsst/cmservice/db/script.py
@@ -19,7 +19,9 @@ from .group import Group
 from .job import Job
 from .node import NodeMixin
 from .row import RowMixin
-from .specification import SpecBlock, SpecBlockAssociation, Specification
+from .spec_block import SpecBlock
+from .spec_block_association import SpecBlockAssociation
+from .specification import Specification
 from .step import Step
 
 if TYPE_CHECKING:
